Move FieldFilter to smithy4s.schema
