+++ modules/core/src/smithy4s/Document.scala
@@ -21,7 +21,7 @@ import smithy4s.schema.CachedSchemaCompiler
 import internals.DocumentDecoderSchemaVisitor
 import internals.DocumentEncoderSchemaVisitor
 import smithy4s.codecs.PayloadError
-import smithy4s.codecs.FieldFilter
+import smithy4s.schema.FieldFilter
 
 /**
   * A json-like free-form structure serving as a model for
